+++ mountpoint-s3-fs/examples/fs_benchmark.rs
@@ -174,6 +174,7 @@ fn mount_file_system(
         SuperblockConfig {
             cache_config: filesystem_config.cache_config.clone(),
             s3_personality: filesystem_config.s3_personality,
+            allow_copy_rename: filesystem_config.allow_copy_rename,
         },
     );
     let fs = S3Filesystem::new(client, prefetcher_builder, pool, runtime, superblock, filesystem_config);